backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
ultralytics
anthropic
python-dotenv
sqlalchemy
pillow
numpy
python-multipart
groq
